Setup Spotlight to capture dev exceptions
This sets up https://spotlightjs.com/ catch exceptions that would otherwise go up to Sentry in development.

Additionally the Sentry exception handling in development is disabled and now depends on DSNs that are baked into/supplied for the images rather than a DSN that is hard coded. This also removed a DSN that is public and new Sentry DSNs have been generated.