Redirects for tide gauges
This sets up URLs using Next.js redirect functionality to provide slightly shorter URLs that more importantly are stable (and can be adapted if the naming of the gauges changes, or they should be pointed towards the water level pages instead). 

This should be deployed to the main site before the signs are installed.

Closes #3011