Sentry and some other dev updates
Updated Sentry to v8 and trying to enable sourcemaps to make errors a little more understandable.

Added a global-error to try to make things nicer when it fully explodes.

Also removed yarn.lock as it was confusing some tools such as Sentry's wizard now that we've switched back to NPM.