Global loading status
This uses https://github.com/Skyleen77/next-nprogress-bar to put a loading status line across the top (in buoy yellow) during transitions between pages. 

The native Next.js way tends to hide parts of the page when they are being replaced, so this is a different way of showing that status change.